v1.0.2

+++ commands/multiplevoting.js
@@ -8,9 +8,11 @@ module.exports.run = async (client, message, args) => {
 	}
 	
 	const dil = db.fetch(`dil_${message.guildID}`) || "english";
+	const prefix = db.fetch(`prefix_${message.guildID}`) || ".";
 	
 	if (dil == "english") {
 		if (!message.member.permissions.has('administrator')) return message.channel.createMessage(`You must have Administrator permission to use this command.`)
+		if (db.has(`denyvoting_${message.guildID}`)) return message.channel.createMessage(`You must allow voting in suggestions with \`${prefix}allowvoting\` comand before opening this feature.`)
 		if (db.has(`multiplevoting_${message.guildID}`)) {
 			db.delete(`multiplevoting_${message.guildID}`)
 			message.channel.createMessage(`Successfully allowed multiple voting.`)
@@ -22,6 +24,7 @@ module.exports.run = async (client, message, args) => {
 	
 	if (dil == "turkish") {
 		if (!message.member.permissions.has('administrator')) return message.channel.createMessage(`Bu komudu kullanmak için Yönetici yetkisine sahip olmalısın.`)
+		if (db.has(`denyvoting_${message.guildID}`)) return message.channel.createMessage(`Bu özelliği açmadan önce \`${prefix}oylamaizni\` komuduyla önerilerdeki oylamayı açmalısın.`)
 		if (db.has(`multiplevoting_${message.guildID}`)) {
 			db.delete(`multiplevoting_${message.guildID}`)
 			message.channel.createMessage(`Başarıyla önerilerde birden fazla tepkiye oy verebilme açıldı.`)
